backend/requirements.txt
django==4.1.7
channels==4.0.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
django-cors-headers==3.13.0
ipdb==0.13.11
daphne==4.0.0
dj-rest-auth==3.0.0
django-allauth==0.52.0
djangorestframework-simplejwt==5.2.2